requirements.txt
Django>=5.0
djangorestframework>=3.14
mysqlclient>=2.2
python-dotenv>=1.0
